+++ api/lib/MqttClient.ts
@@ -7,7 +7,13 @@ import {
 	IClientSubscribeOptions,
 	connect,
 } from 'mqtt'
-import { allSettled, parseJSON, sanitizeTopic, pkgJson } from './utils'
+import {
+	allSettled,
+	parseJSON,
+	sanitizeTopic,
+	pkgJson,
+	stringifyJSON,
+} from './utils'
 import { module } from './logger'
 import { TypedEventEmitter } from './EventEmitter'
 import { storeDir } from '../config/app'
@@ -311,7 +317,7 @@ class MqttClient extends TypedEventEmitter<MqttClientEventCallbacks> {
 
 			this.client.publish(
 				topic,
-				JSON.stringify(data),
+				stringifyJSON(data),
 				options,
 				function (err) {
 					if (err) {
@@ -529,7 +535,10 @@ class MqttClient extends TypedEventEmitter<MqttClientEventCallbacks> {
 
 		// It's an action
 		if (parts[0] === MqttClient.CLIENTS_PREFIX) {
-			if (parts.length < 3) return
+			if (parts.length < 3 || parts[1] !== this._clientID) {
+				// it could be we receive a message from another Z-UI client, ignore it
+				return
+			}
 
 			const action = MqttClient.ACTIONS.indexOf(parts[2])