+++ src/components/Confirm.vue
@@ -225,7 +225,7 @@
 </template>
 
 <script>
-import { tryParseDSKFromQRCodeString } from '@zwave-js/core/safe'
+import { tryParseDSKFromQRCodeString } from '@zwave-js/core'
 import 'vue-prism-editor/dist/prismeditor.min.css' // import the styles somewhere
 
 // import highlighting library (you can use any library you want just return html string)