+++ src/components/nodes-table/NodeScheduler.vue
@@ -83,8 +83,8 @@ import { mapActions } from 'pinia'
 
 import useBaseStore from '../../stores/base.js'
 import InstancesMixin from '../../mixins/InstancesMixin.js'
-import { getEnumMemberName } from 'zwave-js/safe'
-import { ScheduleEntryLockWeekday } from '@zwave-js/cc/safe'
+import { getEnumMemberName } from '@zwave-js/shared'
+import { ScheduleEntryLockWeekday } from '@zwave-js/cc'
 import { padNumber, copy } from '../../lib/utils.js'
 
 const months = [